Adds precision method to DatetimeComponents
An HL7 DTM value specifies its precision by its length. For example,
while "1990" and "19900101" might be parsed as the same timestamp,
the former indicates that it is precise to the year, while the
second indicates that it is precise to the month. This is a
meaningful distinction that is lost by simply producing date/time
values from the two strings.

This PR adds a method, DatetimeComponents#precision, which
returns the name of the least significant component that is
present in the input. The possible return values are:

year
month
day
hour
minute
second
fraction

Note that nil is not a possible return value, since all DTM values
must provide at least a year to be valid. Trying to create a
DatetimeComponent value from an invalid DTM string raises an
exception.